vllm: Use a cuda base container image
I ran into errors when trying to use multi-gpu support in vllm because
libcuda was not in the image. The failure presented as NCCL failing to
initialize. Switch to a cuda base image so more cuda dependencies are
in place.

I never actually got this to work completely, but didn't want to throw
away the change just yet and it doesn't seem to break anything for the
single GPU case at the moment.

Signed-off-by: Russell Bryant <rbryant@redhat.com>